Refactoring involving renaming, code clean up, comments, a tad jQuery.
Some renaming happened: 
	— FirstController —> AccountController
	— index.jsp (used to be a form to create a new account, so I renamed
it…) —> createAccount.jsp

Refactored the “FirstController” to be called “AccountController” — it
handles all of the interaction with Accounts.
	— added a bunch of comments to keep things clear.
	— Changed the Account class’ mutator methods to use Integer instead of
just an int, so that it would play
		nicely with the AccountRepository.
	— Switched the mapping of “/“ to deliver the wishlist.jsp instead of
the createAccount jsp.

Wrote tons of comments so people won’t forget what’s going on in two
weeks!

Moved the newWishListButton’s o’clock show/hide of the wishListEditDiv
into the $(document).ready () of the wishlist.jsp (it was previously in
the wishlist.js).
	— Also, added a hide() to the #wishListAdmin div, so that clicking
Create New WishList or clicking on an existing one
		hides the admin panel in either case.

Added some handling of errors and logout messages on the login.jsp page.
Used some jQuery to hide any messages
after 2.5 seconds of being visible. Used the Window.setTimeout function
for the delay because jQuery’s delay() function is 
best used in between jQuery transitions and often has issues if it is
not used in this way.